create if else
Task
Given an integer, , perform the following conditional actions:

If  n odd, print Weird
If  n is even and in the inclusive range of 2 to 5, print Not Weird
If n  is even and in the inclusive range of  6 to 20 , print Weird
If n  is even and greater than 20 , print Not Weird
Input Format

A single line containing a positive integer, .

Constraints
1<=n<=100
Output Format

Print Weird if the number is weird. Otherwise, print Not Weird.